move delay pattern to class
